requirements.txt
python-telegram-bot==21.6
yt-dlp==2025.1.12
aiohttp==3.10.5
